Fjernet tjeneste for å hente grunnlag sekvensielt (hentes nå asynkront). Fjernet noe logging.
