+++ src/main/kotlin/no/nav/bidrag/grunnlag/service/HentBarnetilsynService.kt
@@ -8,22 +8,12 @@ import no.nav.bidrag.grunnlag.consumer.familieefsak.api.BarnetilsynRequest
 import no.nav.bidrag.grunnlag.consumer.familieefsak.api.BarnetilsynResponse
 import no.nav.bidrag.grunnlag.exception.RestResponse
 import no.nav.bidrag.transport.behandling.grunnlag.response.BarnetilsynGrunnlagDto
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
 
-// @Service
 class HentBarnetilsynService(
     private val familieEfSakConsumer: FamilieEfSakConsumer,
 ) : List<BarnetilsynGrunnlagDto> by listOf() {
 
-    companion object {
-        @JvmStatic
-        val LOGGER: Logger = LoggerFactory.getLogger(HentBarnetilsynService::class.java)
-    }
-
     fun hentBarnetilsyn(barnetilsynRequestListe: List<PersonIdOgPeriodeRequest>): List<BarnetilsynGrunnlagDto> {
-        LOGGER.info("Start BARNETILSYN")
-
         val barnetilsynListe = mutableListOf<BarnetilsynGrunnlagDto>()
 
         barnetilsynRequestListe.forEach {
@@ -36,17 +26,16 @@ class HentBarnetilsynService(
                 val restResponseBarnetilsyn = familieEfSakConsumer.hentBarnetilsyn(hentBarnetilsynRequest)
             ) {
                 is RestResponse.Success -> {
-                    SECURE_LOGGER.info("Barnetilsyn enslig forsørger ga følgende respons: ${restResponseBarnetilsyn.body}")
+                    SECURE_LOGGER.info("Henting av barnetilsyn ga følgende respons for ${it.personId}: ${restResponseBarnetilsyn.body}")
                     leggTilBarnetilsyn(barnetilsynListe, restResponseBarnetilsyn.body, it.personId)
                 }
 
                 is RestResponse.Failure -> {
-                    SECURE_LOGGER.warn("Feil ved henting av barnetilsyn enslig forsørger for ${it.personId}")
+                    SECURE_LOGGER.warn("Feil ved henting av barnetilsyn for ${it.personId}")
                 }
             }
         }
 
-        LOGGER.info("Slutt BARNETILSYN")
         return barnetilsynListe
     }
