+++ src/main/kotlin/no/nav/bidrag/grunnlag/service/HentSivilstandService.kt
@@ -6,22 +6,12 @@ import no.nav.bidrag.grunnlag.consumer.bidragperson.BidragPersonConsumer
 import no.nav.bidrag.grunnlag.exception.RestResponse
 import no.nav.bidrag.transport.behandling.grunnlag.response.SivilstandGrunnlagDto
 import no.nav.bidrag.transport.person.SivilstandPdlHistorikkDto
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
 
-// @Service
 class HentSivilstandService(
     private val bidragPersonConsumer: BidragPersonConsumer,
 ) : List<SivilstandGrunnlagDto> by listOf() {
 
-    companion object {
-        @JvmStatic
-        val LOGGER: Logger = LoggerFactory.getLogger(HentSivilstandService::class.java)
-    }
-
     fun hentSivilstand(sivilstandRequestListe: List<PersonIdOgPeriodeRequest>): List<SivilstandGrunnlagDto> {
-        LOGGER.info("Start SIVILSTAND")
-
         val sivilstandListe = mutableListOf<SivilstandGrunnlagDto>()
 
         sivilstandRequestListe.forEach {
@@ -30,7 +20,7 @@ class HentSivilstandService(
             ) {
                 is RestResponse.Success -> {
                     SECURE_LOGGER.info(
-                        "Kall til bidrag-person for å hente sivilstand ga følgende respons for ${it.personId}: ${restResponseSivilstand.body}",
+                        "Henting av sivilstand sivilstand ga følgende respons for ${it.personId}: ${restResponseSivilstand.body}",
                     )
                     leggTilSivilstand(sivilstandListe, restResponseSivilstand.body, it.personId)
                 }
@@ -41,7 +31,6 @@ class HentSivilstandService(
             }
         }
 
-        LOGGER.info("Slutt SIVILSTAND")
         return sivilstandListe
     }
