+++ src/main/kotlin/no/nav/bidrag/grunnlag/service/HentUtvidetBarnetrygdOgSmåbarnstilleggService.kt
@@ -6,24 +6,14 @@ import no.nav.bidrag.grunnlag.consumer.familiebasak.api.FamilieBaSakRequest
 import no.nav.bidrag.grunnlag.consumer.familiebasak.api.FamilieBaSakResponse
 import no.nav.bidrag.grunnlag.exception.RestResponse
 import no.nav.bidrag.transport.behandling.grunnlag.response.UtvidetBarnetrygdOgSmaabarnstilleggGrunnlagDto
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
 import java.math.BigDecimal
 import java.time.LocalDate
 
-// @Service
 class HentUtvidetBarnetrygdOgSmåbarnstilleggService(
     private val familieBaSakConsumer: FamilieBaSakConsumer,
 ) : List<UtvidetBarnetrygdOgSmaabarnstilleggGrunnlagDto> by listOf() {
 
-    companion object {
-        @JvmStatic
-        val LOGGER: Logger = LoggerFactory.getLogger(HentUtvidetBarnetrygdOgSmåbarnstilleggService::class.java)
-    }
-
     fun hentUbst(ubstRequestListe: List<PersonIdOgPeriodeRequest>): List<UtvidetBarnetrygdOgSmaabarnstilleggGrunnlagDto> {
-        LOGGER.info("Start UTVIDET_BARNETRYGD_OG_SMÅBARNSTILLEGG")
-
         val ubstListe = mutableListOf<UtvidetBarnetrygdOgSmaabarnstilleggGrunnlagDto>()
 
         ubstRequestListe.forEach {
@@ -37,7 +27,7 @@ class HentUtvidetBarnetrygdOgSmåbarnstilleggService(
             ) {
                 is RestResponse.Success -> {
                     SECURE_LOGGER.info(
-                        "Familie-BA-sak hent utvidet barnetrygd og småbarnstillegg ga følgende respons: ${restResponseUbst.body}",
+                        "Henting av utvidet barnetrygd og småbarnstillegg ga følgende respons for ${it.personId}: ${restResponseUbst.body}",
                     )
                     leggTilUbst(ubstListe, restResponseUbst.body, it.personId)
                 }
@@ -48,7 +38,6 @@ class HentUtvidetBarnetrygdOgSmåbarnstilleggService(
             }
         }
 
-        LOGGER.info("Slutt UTVIDET_BARNETRYGD_OG_SMÅBARNSTILLEGG")
         return ubstListe
     }
